refactor(api): modify structure of files to a new organized API
