+++ Backend/docs/swaggerConfig.js
@@ -1,6 +1,6 @@
 const swaggerUi = require('swagger-ui-express')
 const YAML = require('yamljs')
-const swaggerDocument = YAML.load('./APIdocs/api-spec.yaml')
+const swaggerDocument = YAML.load('./docs/api-spec.yaml')
 
 const setupSwagger = (app) => {
   app.use(
